# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/candidates/me": {
      "get": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Get own profile",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Update profile fields",
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/candidates/me/verify-identity": {
      "post": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Initiate RENAPER verification",
        "operationId": "verifyIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/me/projects": {
      "post": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Add project",
        "operationId": "addProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/candidates/me/photo": {
      "post": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Upload profile photo",
        "operationId": "uploadPhoto",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/me/experience": {
      "post": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Add work experience",
        "operationId": "addExperience",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkExperienceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register user",
        "description": "Registers a new Candidate or Recruiter",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User login",
        "description": "Authenticates a user and returns a JWT",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/anonymous-threads": {
      "post": {
        "tags": [
          "Anonymous Messaging"
        ],
        "summary": "Candidate creates anonymous thread",
        "operationId": "createThread",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThreadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/anonymous-threads/{id}/messages": {
      "post": {
        "tags": [
          "Anonymous Messaging"
        ],
        "summary": "Send a message within an existing thread",
        "operationId": "sendMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/offers/{offerId}/anonymous-threads": {
      "get": {
        "tags": [
          "Anonymous Messaging"
        ],
        "summary": "Recruiter sees anonymous threads for their offer",
        "operationId": "getOfferThreads",
        "parameters": [
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/me/projects/{id}": {
      "delete": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Remove project",
        "operationId": "removeProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/candidates/me/experience/{id}": {
      "delete": {
        "tags": [
          "Candidate Profile"
        ],
        "summary": "Remove work experience",
        "operationId": "removeExperience",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string"
          },
          "currentRole": {
            "type": "string"
          },
          "headline": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "linkedIn": {
            "type": "string"
          }
        }
      },
...
      "VerifyIdentityRequest": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string"
          },
          "tramiteNumber": {
            "type": "string"
          }
        }
      },
...
      "ProjectRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "link": {
            "type": "string"
          }
        }
      },
      "WorkExperienceRequest": {
        "type": "object",
        "properties": {
          "company": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "userType": {
            "type": "string",
            "minLength": 1
          },
          "fullName": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "fullName",
          "password",
          "userType"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          }
        }
      },
      "CreateThreadRequest": {
        "type": "object",
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
    }
  }
}